requirements.txt
torch==2.3.1
torchvision
scipy
numpy
Pillow
mock
scikit-image
imageio
gitpython
lifelines
docutils==0.16.0
markdown>=3.4.0
myst_parser
-e git+https://github.com/Westlake-AI/pytorch_sphinx_theme.git#egg=pytorch_sphinx_theme
recommonmark
sphinx==4.0.2
sphinx-copybutton
sphinx_markdown_tables>=0.0.16
sphinx_rtd_theme==0.5.2
urllib3==1.26.15
basemap
geos
pyproj
tensorflow
dask
decord
future
fvcore
hickle
lpips
matplotlib
nni
netcdf4==1.5.8
opencv-python
packaging
pandas
six
scikit-learn
timm>=0.5.4,<=0.6.11
tqdm
xarray==0.19.0
asynctest
codecov
flake8
isort
pytest 
pytest-cov
pytest-runner
xdoctest >= 0.10.0
yapf
kwarray

